requirements.txt
anyio==3.6.1
APScheduler==3.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.24
dnspython==2.2.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
markovify==0.9.4
pymongo==4.2.0
PySocks==1.7.1
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.5
tzlocal==4.2
Unidecode==1.3.6
